+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `Array::set_shape_and_chunk_grid`
 - Add `node::[async_]get_child_nodes_opt`
+- Add `Group::storage()`
 
 ### Changed
 - Bump `zarrs_metadata_ext` to 0.2.2